requirements.txt
wheel
six
screen-search
pynput
numpy
matplotlib
Pillow
PyAutoGUI
opencv-python